+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 3.3.4 [unreleased]
 
 ### Added
+* [#187](https://github.com/stlehmann/pyads/pull/187) Support structured data types in `read_list_by_name`
 * [#195](https://github.com/stlehmann/pyads/pull/195) Read/write by name without passing the datatype
 * [#200](https://github.com/stlehmann/pyads/pull/200) Split read write by list into max-ads-sub-comands chunks
 * [#206](https://github.com/stlehmann/pyads/pull/206) AdsSymbol now supports DT, DATE_TIME and TIME datatypes