+++ CHANGELOG.md
@@ -3,7 +3,16 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
-## 3.3.4 [unreleased]
+
+## 3.3.5 [unreleased]
+
+### Added
+
+### Changed
+
+### Removed
+
+## 3.3.4
 
 ### Added
 * [#187](https://github.com/stlehmann/pyads/pull/187) Support structured data types in `read_list_by_name`